requirements.txt
pyyaml>=6.0.1
watchdog>=3.0.0
redis>=5.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
requests>=2.31.0
flask>=3.0.0
apscheduler>=3.10.0
ua-parser>=0.18.0
geoip2>=4.7.0
flask-cors>=4.0.0

